+++ EnviosApp/Models/Zone.cs
@@ -2,8 +2,8 @@
 {
     public class Zone
     {
-        public int Id { get; set; }
-        public int ProviderId { get; set; } // FK a Provider
+        public long Id { get; set; }
+        public long ProviderId { get; set; } // FK a Provider
         public string Name { get; set; } // Ejemplo: Zona 1, Zona A
         public decimal BasePrice { get; set; } // Precio base
         // Relación muchos a muchos con Zone
